feat: add preview image to blog list view
