+++ src/utils/blog.ts
@@ -7,6 +7,7 @@ export const BlogMetadata = z.object({
   date: z.string(),
   author: z.string(),
   tags: z.array(z.string()),
+  thumbnail: z.string().optional(),
 });
 
 export type TBlogMetadata = z.infer<typeof BlogMetadata>;
